rename guestRequest.ts to request.ts
