+++ MCA/api/routes/staff/requests.ts
@@ -4,7 +4,7 @@ import { currentMCA } from "../../../../MCA-AYIM/api/middleware";
 import { GuestRequest } from "../../../../Models/MCA_AYIM/guestRequest";
 import { MCA } from "../../../../Models/MCA_AYIM/mca";
 import { MCAEligibility } from "../../../../Models/MCA_AYIM/mcaEligibility";
-import { RequestStatus } from "../../../../Interfaces/guestRequests";
+import { RequestStatus } from "../../../../Interfaces/requests";
 
 const staffRequestsRouter = new Router;
